Separating MG tools in a separate source file
